Complete rework of the OSM geocoding part
The OSM data is converted into a normalized DB structure in the
'optimize'-phase. Additionally we try to merge old style
house-numbering ('addr:*' tag on a building) with the newer style
(separate node that describes the location of the entry to the building)

The geocoder tries OSM data first and if that does not yield a result
it falls back to openaddresses.io data.

Sharding is not handled explicitly anymore because PostgreSQL 11 can
handle it on it's own.